Added Achievements, hobbies, contact gists + UI
Created Gists for above and displayed them in UI. Note: This is
generalized now! It is possible for anyone to add their data, based on
the gists I have used! and get their own version of this portfolio!
Ofcourse, if any bugs, kindly open an issue :).
Next, FIGURE OUT a way to display ```github_archive```. We have the API
setup to extract data using Google BigQuery + GithubArchive.org project